Added some polymorphic behaviour
Made the code more modular using polymorphism.
virtual methods are declared in class User.
Different kinds of users are handled.

+++ Inventory/Inventory.h
@@ -20,7 +20,7 @@ public:
     Product *getReference(const std::string &id);
 
     bool isUniqueId (const std::string& id) { 
-        for (int i=0; i<products.size(); i++) { 
+        for (size_t i=0; i<products.size(); i++) { 
             if (id == products.at(i)->get_unique_id()) { 
                 return true; 
             }
